Improved user_drivers/validate.sh script to loop through folders and check if the folder has a validate.sh script, then run that script.

+++ validate.sh
@@ -1,32 +1,42 @@
 #!/bin/bash
 
-CURRENTDIR=`pwd`
-folders=("lagrange_square" \
-"lagrange_step")
+USER_DRIVERS_DIR=`pwd`
+#folders=("lagrange_square" \
+#"lagrange_step")
 
 touch validation.log
 rm -rf validation.log
 touch validation.log
 
-for i in "${folders[@]}"
+for d in */ ;
 do
-  echo "Doing $i" >> validation.log
-  cd $i
-  ./validate.sh 2>&1 | tee validate.output
-  cat ./Validate/validation.log >> $CURRENTDIR/validation.log
-
-  ## Now, if the ./Validate/validation.log file is empty, we know that
-  ## all the tests pass, thus we can safely delete all the associated files
-  ## of this self test.
-  if [ ! -s ./Validate/validation.log ]
+  if [[ -d "$d" && -f "$d/validate.sh" ]];
   then
-    rm -rf Validate validate.output
-  fi
+    echo "$d/Validate/validate.log:" >> validation.log
 
-  cd $CURRENTDIR
-done
+    ## Go into the directory
+    cd $d
 
-cd $CURRENTDIR
+    ## Run the validate script
+    ./validate.sh 2>&1 | tee validate.output
+
+    ## Now, if the ./Validate/validation.log file is empty, we know that
+    ## all the tests pass, thus we can safely delete all the associated files
+    ## of this self test. -s is true if file is not zero size.
+    if [ -s ./Validate/validation.log ]
+    then
+      cat ./Validate/validation.log >> $USER_DRIVERS_DIR/validation.log
+    else
+      echo "OK" >> $USER_DRIVERS_DIR/validation.log
+      rm -rf Validate validate.output
+    fi
+  
+    cd $USER_DRIVERS_DIR
+
+  fi # If the folder contains a validate.sh script.
+done # for loop through folders.
+
+cd $USER_DRIVERS_DIR
 cat validation.log
